pyproject.toml
[project]
name = "multi-agentapp-using-adk"
version = "1.0"
description = "multi agent app using google adk with custom agent"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "google-adk>=1.2.1",
    "litellm>=1.72.1",
]
